Fix compiler warnings

+++ src/options.cc
@@ -367,7 +367,7 @@ void COptions::Parse(int argc, char *argv[])
 
       // Create name stripped from any cross/ or native/ parts
       Config.name_stripped = Config.name;
-      int pos = Config.name_stripped.rfind("/");
+      size_t pos = Config.name_stripped.rfind("/");
       if (pos != string::npos)
          Config.name_stripped.erase(0,++pos);
